Add support for per-metric tags
Tag values are extracted from the metric name using capture groups defined
as part of the measurement mapping.  Unfortunately there is no public way
to find out what groups are defined in a pattern so we have to declare the
expected tag keys explicitly.

The default behavior is that the value for each tag will be extracted by
looking for a capture group named with the tag key.  Unfortunately Java
also doesn't support duplicate group names even in mutually exclusive
sections of the regex, so to handle those cases we have to allow for the
definition of a custom tag extraction function which will be called to produce
the value for the given key.

All of this is a bit messy, but there aren't many alternatives (at least
until Dropwizard supports attaching tags to the metric names directly).